Move `krate::metadata::versions` to `krate:versions::versions`
Prepare for implementing pagination support for the versions endpoint.
